Introduce new 'naver-map-location' artifact
