+++ gradle.properties
@@ -27,9 +27,6 @@ SONATYPE_HOST=S01
 RELEASE_SIGNING_ENABLED=true
 
 GROUP=io.github.fornewid
-VERSION_NAME=1.5.1-SNAPSHOT
-
-POM_DESCRIPTION=NAVER Map Android SDK for Jetpack Compose
 
 POM_URL=https://github.com/fornewid/naver-map-compose/
 POM_SCM_URL=https://github.com/fornewid/naver-map-compose/
