+++ naver-map-compose/gradle.properties
@@ -1,3 +1,5 @@
 POM_ARTIFACT_ID=naver-map-compose
 POM_NAME=NAVER Map Android SDK for Jetpack Compose
+POM_DESCRIPTION=NAVER Map Android SDK for Jetpack Compose
 POM_PACKAGING=aar
+VERSION_NAME=1.5.1-SNAPSHOT
