+++ naver-map-location/api.txt
@@ -1,0 +1,15 @@
+// Signature format: 4.0
+package com.naver.maps.map.location {
+
+  public abstract class FusedLocationSource implements com.naver.maps.map.LocationSource {
+    ctor public FusedLocationSource(android.content.Context context);
+    method public void activate(com.naver.maps.map.LocationSource.OnLocationChangedListener listener);
+    method public void deactivate();
+    method public abstract boolean hasPermissions();
+    method public final void onPermissionGranted();
+    method public abstract void onPermissionRequest();
+    method public final void setCompassEnabled(boolean enabled);
+  }
+
+}
+
