+++ naver-map-location/build.gradle
@@ -1,13 +1,13 @@
 plugins {
     id 'com.android.library'
-    id 'kotlin-android'
+    id 'org.jetbrains.kotlin.android'
     id 'com.vanniktech.maven.publish'
     id 'org.jetbrains.dokka'
     id 'me.tylerbwong.gradle.metalava'
 }
 
 android {
-    namespace 'com.naver.maps.map.compose'
+    namespace 'com.naver.maps.map.location'
     compileSdkVersion buildConfig.compileSdk
     defaultConfig {
         minSdkVersion buildConfig.minSdk
@@ -19,35 +19,16 @@ android {
         sourceCompatibility JavaVersion.VERSION_1_8
         targetCompatibility JavaVersion.VERSION_1_8
     }
-    composeOptions {
-        kotlinCompilerExtensionVersion "${libs.versions.composeCompiler.get()}"
-    }
     buildFeatures {
         buildConfig false
-        compose true
     }
     kotlinOptions {
         freeCompilerArgs += '-Xexplicit-api=warning'
     }
 }
 
-metalava {
-    sourcePaths.setFrom("src/main")
-    filename.set("api/current.api")
-    reportLintsAsErrors.set(true)
-}
-
 dependencies {
-    api libs.naver.map
-    api project(':naver-map-location')
-
-    implementation libs.compose.foundation
+    implementation libs.naver.map
     implementation libs.kotlin.stdlib
-
-    // To provide FusedLocationSource
     implementation libs.google.play.services.location
-    implementation libs.accompanist.permissions
-
-    androidTestImplementation libs.androidx.test.ext.junit
-    androidTestImplementation libs.androidx.test.espresso.core
 }
