+++ naver-map-location/gradle.properties
@@ -1,0 +1,5 @@
+POM_ARTIFACT_ID=naver-map-location
+POM_NAME=Location for NAVER Map Android SDK
+POM_DESCRIPTION=Location for NAVER Map Android SDK
+POM_PACKAGING=aar
+VERSION_NAME=16.0.0-SNAPSHOT
